+++ packages/devtools_shared/lib/src/extensions/extension_model.dart
@@ -13,6 +13,7 @@ class DevToolsExtensionConfig implements Comparable {
     required this.issueTrackerLink,
     required this.version,
     required this.materialIconCodePoint,
+    required this.isPubliclyHosted,
   });
 
   factory DevToolsExtensionConfig.parse(Map<String, Object?> json) {
@@ -33,6 +34,7 @@ class DevToolsExtensionConfig implements Comparable {
           pathKey: final String path,
           issueTrackerKey: final String issueTracker,
           versionKey: final String version,
+          isPubliclyHostedKey: final String isPubliclyHosted,
         }) {
       return DevToolsExtensionConfig._(
         name: name,
@@ -40,9 +42,16 @@ class DevToolsExtensionConfig implements Comparable {
         issueTrackerLink: issueTracker,
         version: version,
         materialIconCodePoint: codePoint,
+        isPubliclyHosted: bool.parse(isPubliclyHosted),
       );
     } else {
-      const requiredKeys = {nameKey, pathKey, issueTrackerKey, versionKey};
+      const requiredKeys = {
+        nameKey,
+        pathKey,
+        issueTrackerKey,
+        versionKey,
+        isPubliclyHostedKey,
+      };
       final diff = requiredKeys.difference(json.keys.toSet());
       if (diff.isEmpty) {
         // All the required keys are present, but the value types did not match.
@@ -71,6 +80,7 @@ class DevToolsExtensionConfig implements Comparable {
   static const issueTrackerKey = 'issueTracker';
   static const versionKey = 'version';
   static const materialIconCodePointKey = 'materialIconCodePoint';
+  static const isPubliclyHostedKey = 'isPubliclyHosted';
 
   /// The package name that this extension is for.
   final String name;
@@ -104,7 +114,12 @@ class DevToolsExtensionConfig implements Comparable {
   /// See https://github.com/flutter/flutter/blob/master/packages/flutter/lib/src/material/icons.dart.
   final int materialIconCodePoint;
 
+  /// Whether this extension is distrubuted in a public package on pub.dev.
+  final bool isPubliclyHosted;
+
   String get displayName => name.toLowerCase();
+
+  String get analyticsSafeName => isPubliclyHosted ? name : 'private';
 
   Map<String, Object?> toJson() => {
         nameKey: name,
@@ -112,6 +127,7 @@ class DevToolsExtensionConfig implements Comparable {
         issueTrackerKey: issueTrackerLink,
         versionKey: version,
         materialIconCodePointKey: materialIconCodePoint,
+        isPubliclyHostedKey: isPubliclyHosted.toString(),
       };
 
   @override
@@ -132,7 +148,8 @@ class DevToolsExtensionConfig implements Comparable {
         other.path == path &&
         other.issueTrackerLink == issueTrackerLink &&
         other.version == version &&
-        other.materialIconCodePoint == materialIconCodePoint;
+        other.materialIconCodePoint == materialIconCodePoint &&
+        other.isPubliclyHosted == isPubliclyHosted;
   }
 
   @override
@@ -142,6 +159,7 @@ class DevToolsExtensionConfig implements Comparable {
         issueTrackerLink,
         version,
         materialIconCodePoint,
+        isPubliclyHosted,
       );
 }