+++ packages/devtools_extensions/CHANGELOG.md
@@ -7,7 +7,7 @@ DevTools extension.
 * Add an example `launch.json` file in the `example/foo` directory.
 * Clean up the package readme to make instructions Windows-compatible.
 * Update the README with instructions for joining the Flutter Discord server.
-* Bump `package:devtools_shared` dependency to ^6.0.0
+* Bump `package:devtools_shared` dependency to ^6.0.1
 * Bump `package:devtools_app_shared` dependency to ^0.0.7
 * Bump `package:vm_service` dependency to ^13.0.0.