+++ packages/devtools_shared/CHANGELOG.md
@@ -1,3 +1,6 @@
+# 6.0.1
+* Add field `isPublic` to `DevToolsExtensionConfig`.
+
 # 6.0.0
 * Bump `package:vm_service` dependency to ^13.0.0.
 * Remove `ServiceCreator` typedef and replace usages with `VmServiceFactory` typedef from `package:vm_service`.